+++ scripts/macros/spells/hailOfThorns.js
@@ -18,7 +18,7 @@ export async function hailOfThorns({speaker, actor, token, character, item, args
     if (!originItem) return;
     featureData.system.save.dc = chris.getSpellDC(originItem);
     setProperty(featureData, 'chris-premades.spell.castData.school', originItem.system.school);
-    let feature = new CONFIG.Item.documentClass(featureData, {parent: workflow.actor});
+    let feature = new CONFIG.Item.documentClass(featureData, {'parent': workflow.actor});
     let targetToken = workflow.targets.first();
     let targetUuids = await chris.findNearby(targetToken, 5).concat(targetToken).map(t=>t.document.uuid);
     let options = {